memory and stack are consistently working on unsigned 16 bit values now
git-svn-id: https://zmpp.svn.sourceforge.net/svnroot/zmpp/trunk@650 7df28f57-6730-0410-8a28-a3245024edd1

+++ zmpp/src/java/org/zmpp/vm/Dictionary.java
@@ -54,10 +54,9 @@ public interface Dictionary {
   
   /**
    * Returns the number of dictionary entries.
-   * 
    * @return the number of entries
    */
-  int getNumberOfEntries();
+  short getNumberOfEntries();
   
   /**
    * Returns the entry address at the specified position.
